Added a helper function run_in_transaction, also fixed a bug in docker-version of the homework
As suggested by Mr. Riccardo Tommasini I added a helper function from my homework to the sql homework template.

It adds the above mentioned function that runs a provided query in a transaction, committing it or rolling it back in the end. This will greatly reduce the amount of boilderplate code that students need to write.

I also fixed an oversight in the docker version of the homework where instead of installing psycopg2 the cell checked whether it is installed. In a fresh docker container it is not installed, so it makes sense to replace it with "! pip install psycopg2"